ci/requirements.txt
apache-beam==2.24.0
attrs==20.3.0
avro-python3==1.9.2.1
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==0.23.6
fasteners==0.15
future==0.18.2
google-api-core==1.22.4
google-apitools==0.5.31
google-auth==1.22.1
google-cloud-bigquery==1.28.0
google-cloud-bigtable==1.5.1
google-cloud-core==1.4.3
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-spanner==1.19.1
google-cloud-storage==1.32.0
google-cloud-videointelligence==1.16.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
grpcio-gcp==0.2.2
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
mock==2.0.0
monotonic==1.5
numpy==1.19.2
oauth2client==3.0.0
packaging==20.4
pandas==1.1.3
pbr==5.5.1
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.11
zipp==3.4.0
